requirements.txt
pillow>=6.2.0
onnxruntime>=1.3.0
numpy>=1.17.2